Cast to eliminate compiler warning

+++ src/Core/Containers/StaticArray.h
@@ -67,7 +67,7 @@ public:
 
   T& operator[](int index)
   {
-    CHECKARRAYBOUNDS(index, 0, size_);
+    CHECKARRAYBOUNDS(index, 0, (int)size_);
     return data_[index];
   }
